🎨 added more functionality to the code

+++ MongoDBBloggerPost/Controller/PostsController.cs
@@ -3,6 +3,7 @@ using System.Collections.Generic;
 using System.Linq;
 using System.Threading.Tasks;
 using Microsoft.AspNetCore.Mvc;
+using MongoDB.Bson;
 using MongoDBBloggerPost.Core.Services;
 using MongoDBBloggerPost.Model;
 
@@ -25,13 +26,29 @@ namespace MongoDBBloggerPost.Controller
         [HttpGet("GetPost")]
         public async Task<PostsModel> GetPost(string id)
         {
-            return await _postService.GetById(id);
+            try
+            {
+                return await _postService.GetById(id);
+            }
+            catch (System.Exception ex)
+            {
+                System.Console.WriteLine("something went wrong while getting post" + ex.Message);
+                throw;
+            }
         }
 
         [HttpGet("GetAllPosts")]
         public async Task<List<PostsModel>> GetAllPosts()
         {
-            return await _postService.GetAll();
+            try
+            {
+                return await _postService.GetAll();
+            }
+            catch (System.Exception ex)
+            {
+                System.Console.WriteLine("something went wrong while getting all posts" + ex.Message);
+                throw;
+            }
         }
 
         [HttpGet("GetAllPostComments")]
@@ -41,12 +58,14 @@ namespace MongoDBBloggerPost.Controller
             {
                 var post = await _postService.GetById(id);
                 var comments = new List<CommentsModel>();
-                if (post.commentIds != null)
+                if (post.commentIds == null)
                 {
-                    foreach (var commentId in post.commentIds)
-                    {
-                        comments.Add(await _commentService.GetById(commentId.ToString()));
-                    }
+                    throw new ArgumentNullException(nameof(post.commentIds));
+                }
+                
+                foreach (var commentId in post.commentIds)
+                {
+                    comments.Add(await _commentService.GetById(commentId.ToString()));
                 }
                 return comments;
             }
@@ -58,21 +77,60 @@ namespace MongoDBBloggerPost.Controller
         }
 
         [HttpPost("SavePost")]
-        public async Task SavePost(PostsModel post)
+        public async Task SavePost(string userId, PostsModel post)
         {
-            await _postService.Save(post);
+            try
+            {
+                if (post == null)
+                {
+                    throw new ArgumentNullException(nameof(post));
+                }
+                post._id = ObjectId.GenerateNewId();
+                post.id = post._id.ToString();
+                post.userId = userId;
+                await _postService.Save(post);
+            }
+            catch (System.Exception ex)
+            {
+                System.Console.WriteLine("something went wrong while saving post" + ex.Message);
+                throw;
+            }
         }
 
         [HttpPut("UpdatePost")]
         public async Task UpdatePost(PostsModel post)
         {
-            await _postService.Update(post);
+            try
+            {
+                if (post == null)
+                {
+                    throw new ArgumentNullException(nameof(post));
+                }
+                await _postService.Update(post);
+            }
+            catch (System.Exception ex)
+            {
+                System.Console.WriteLine("something went wrong while updating post" + ex.Message);
+                throw;
+            }
         }
 
         [HttpDelete("DeletePost")]
         public async Task DeletePost(PostsModel post)
         {
-            await _postService.Delete(post);
+            try
+            {
+                if (post == null)
+                {
+                    throw new ArgumentNullException(nameof(post));
+                }
+                await _postService.Delete(post);
+            }
+            catch (System.Exception ex)
+            {
+                System.Console.WriteLine("something went wrong while deleting post" + ex.Message); 
+                throw;
+            }
         }
     }
 }
